+++ R/gcap-package.R
@@ -4,8 +4,16 @@
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
-#' @useDynLib gcap, .registration = TRUE
-#' @importFrom Rcpp sourceCpp
+#' @importFrom data.table .BY
+#' @importFrom data.table .EACHI
+#' @importFrom data.table .GRP
+#' @importFrom data.table .I
 #' @importFrom data.table .N
+#' @importFrom data.table .NGRP
+#' @importFrom data.table .SD
+#' @importFrom data.table :=
+#' @importFrom data.table data.table
+#' @importFrom Rcpp sourceCpp
+#' @useDynLib gcap, .registration = TRUE
 ## usethis namespace: end
 NULL